// src/utils/enums.ts
export enum UserRole {
    Admin = 'admin',
    Customer = 'customer',
}

export enum ProductCategory {
    Electronics = 'Electronics',
    Apparel = 'Apparel',
    Footwear = 'Footwear',
    Books = 'Books',
    HomeGoods = 'Home Goods',
    Sports = 'Sports',
}

export enum SortDirection {
    Ascending = 'asc',
    Descending = 'desc',
}